Move DNS::Zone WHOIS info logic to concern
